Are you kidding me! That's still in there?
This was a broken mesh that I was going to revert to if all else failed. The mesh wasn't unusable, but it had the Smelter's collision and probably some unnecessary tags (I still don't really know how to use NifSkope). Don't know how it is STILL there after all of this time.