Fix sloppy CLWATransferOnEnter code
Not sure what I was doing when I wrote parts of that script. Since scripts get baked into saves and I want to make this a drop-in, I've taken the liberty of renaming the script to have an `_` at the end and creating an empty script with the old name.