3.0.2 FOMOD Fix
But the problem is, he forgot to update some file paths in his installer.
I assume that fixing that is going to be very difficult?
Actually, it's going to be super easy—barely an inconvenience!
On really?
Yeah, he just selects them all with his milti-cursor in VS Code and types "\Vanilla"—it's going to take more work for him to actually upload the file to Nexus.
Huh.